iprt/cpuset.h: Shut up sign conversion warnings.  bugref:9790
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@85225 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/iprt/cpuset.h
@@ -328,7 +328,7 @@ DECLINLINE(int) RTCpuLastIndex(PCRTCPUSET pSet)
                     break;
                 u64 <<= 1;
             }
-            return (int)i * 64 + iBit;
+            return (int)i * 64 + (int)iBit;
         }
     }
     return 0;
